refactor: consistent naming and imports for hooks
