+++ src/pages/users/mutations/useUpdateUser.ts
@@ -1,4 +1,4 @@
-import { useMutation } from "hooks/useMutation"
+import { useMutation } from "hooks"
 import { graphql } from "react-relay"
 import { useUpdateUserMutation } from "__generated__/useUpdateUserMutation.graphql"
